refactor(threads): useThreadMessaging 主干外移 squadRequest 协作发送段（runSquadRequestSend，322 行）
